+++ lib/glimmer/tk/menu_proxy.rb
@@ -50,6 +50,10 @@ module Glimmer
         label == 'Window'
       end
       
+      def system?
+        label == 'System'
+      end
+      
       def application?
         @args.first == :application
       end
@@ -63,10 +67,12 @@ module Glimmer
             @parent_proxy.tk.add :cascade, :menu => @tk
           end
         else
-          if @parent_proxy.parent_proxy.is_a?(ToplevelProxy) && OS.mac? && help?
+          if @parent_proxy.parent_proxy.is_a?(ToplevelProxy) && (OS.mac? || OS.linux?) && help?
             @tk = ::TkSysMenu_Help.new(@parent_proxy.tk)
           elsif @parent_proxy.parent_proxy.is_a?(ToplevelProxy) && OS.mac? && window?
             @tk = ::Tk::TkSysMenu_Window.new(@parent_proxy.tk)
+          elsif @parent_proxy.parent_proxy.is_a?(ToplevelProxy) && OS.windows? && system?
+            @tk = ::TkSysMenu_System.new(@parent_proxy.tk)
           else
             tk_widget_class = self.class.tk_widget_class_for(@keyword)
             @tk = tk_widget_class.new(@parent_proxy.tk)